Found the criteria to uniquely list entities of a solid
The TopoDS_Shape::IsSame allows us to compare two entities. When topo
explorer gives us same entities multiple times, we can use this criteria
to check if we have already visited the entity before or not. We record
each newly found entity in a ledger and for each new entity, check if
it's present in the ledger.